+++ AddOns/Decursive/WhatsNew.md
@@ -1,5 +1,51 @@
-Decursive 2.7.34
-================
+Decursive 2.8.0-RC1
+===================
+
+
+Decursive 2.8.0-RC1 (2026-02-27)
+--------------------------------
+
+- Compatible with Midnight
+
+
+Although Decursive is now compatible with WoW 12, some features will be missing depending on the dynamic add-on restrictions being applied:
+    - Cure priority order: when auras are secret, Decursive will always assume that the detected aura is your first priority with the following consequences:
+        - The spell cooldown used will be your first spell's cooldown (so it might be wrong if you have several curing spells)
+        - Wrong mouse button click detection will not work (no alert)
+        - Range detection might not work or be wrong (you can set the most appropriate spell as your first priority to control that)
+
+Gone features in Midgnight:
+    - Debuff timers on MUFs (remaining and elapsed) - Decursive will default to the number of stacks instead. You can still mouse over the live-list to see the debuff timers.
+    - Line of sight failed cast detection and unit black-listing. (Decursive was using the combat log to do that)
+    - Bleed debuff detection (aura's spell ids are no longer accessible)
+    - Debuff filtering (may only work in the open world while outside of combat...)
+    - Stealth detection and reporting while in combat
+    - Decursive can now only be triggered by debuffs that WoW APIs tells you can dispel so it's uncertain if items can still be used reliably...
+
+
+Please note that this is an early release and that there still might be issues.
+
+I'd like to give my warm thanks to Bozoweed for their help in making this release possible.
+
+
+Decursive 2.7.36 (2026-01-25)
+-----------------------------
+
+- Fix TOC for midnight pre-patch
+
+- Fix reported Lua error when debuffs are secret (note that Decursive is
+  currently useless in this situation—see the note about midnight below...
+
+
+**MIDNIGHT NOTE**:
+
+ CLEU being gone and APIs returning secret values, there is little hope but the
+ MUFs and spell detection are still working so there might be a way to do
+ something with the debuf types and the curve color secret related APIs to
+ change the MUFs color depending on the type of debuf...
+
+ For now Decursive will just do nothing if the APIs are in secret mode.
+
 
 Decursive 2.7.34 (2026-01-19)
 -----------------------------
@@ -7,11 +53,6 @@ Decursive 2.7.34 (2026-01-19)
 - Compatible with BCC Anniversary
 
 - Make it so it doesn't crash in Midnight but it's mostly useless.
-
-CLEU being gone and APIs returning secret values, there is little hope but the
-MUFs and spell detection are still working so there might be a way to do
-something with the debuf types and the curve color secret related APIs to
-change the MUFs color depending on the type of debuf...
 
 
 Decursive 2.7.32 (2026-01-01)